scraping-bot/requirements.txt
requests
openpyxl
beautifulsoup4
pdfminer.six
psycopg2
datetime